# opt: simplify merge and lookup join filters