… pkg/sql/opt/exec/execbuilder/testdata/distsql_agg
… pkg/sql/opt/exec/execbuilder/testdata/distsql_interleaved_join
… pkg/sql/opt/exec/execbuilder/testdata/join
… pkg/sql/opt/exec/execbuilder/testdata/srfs
… pkg/sql/opt/memo/statistics_builder.go
… pkg/sql/opt/memo/testdata/logprops/lookup-join
… pkg/sql/opt/memo/testdata/stats/lookup-join
… pkg/sql/opt/norm/testdata/rules/combo
… pkg/sql/opt/norm/testdata/rules/decorrelate
… pkg/sql/opt/norm/testdata/rules/join
… pkg/sql/opt/norm/testdata/rules/project
… pkg/sql/opt/norm/testdata/rules/prune_cols
… pkg/sql/opt/norm/testdata/rules/reject_nulls
… pkg/sql/opt/norm/testdata/rules/select
… pkg/sql/opt/norm/testdata/rules/side_effects
… pkg/sql/opt/xform/custom_funcs.go
… pkg/sql/opt/xform/rules/join.opt
… pkg/sql/opt/xform/testdata/external/activerecord
… pkg/sql/opt/xform/testdata/external/customer
… pkg/sql/opt/xform/testdata/external/hibernate
… pkg/sql/opt/xform/testdata/external/liquibase
… pkg/sql/opt/xform/testdata/external/navicat
… pkg/sql/opt/xform/testdata/external/nova
… pkg/sql/opt/xform/testdata/external/tpcc
… pkg/sql/opt/xform/testdata/external/tpcc-no-stats
… pkg/sql/opt/xform/testdata/external/tpch
… pkg/sql/opt/xform/testdata/physprops/ordering
… pkg/sql/opt/xform/testdata/rules/combo
… pkg/sql/opt/xform/testdata/rules/join