Merge and lookup joins implicitly test equality on a set of columns.
The corresponding equalities can be removed from the ON condition.
This should make execution faster, since we are no longer evaluating
the redundant conditions (at the cost of a bit of planning overhead).

Benchmarks results for the tpcc lookup join query:
```
name                               old time/op  new time/op  delta
Phases/tpcc-stock-level/Parse      2.07ns ± 0%  2.07ns ± 1%    ~     (p=0.794 n=4+5)
Phases/tpcc-stock-level/OptBuild   49.7µs ± 1%  49.9µs ± 2%    ~     (p=0.548 n=5+5)
Phases/tpcc-stock-level/Normalize  50.2µs ± 1%  50.5µs ± 1%    ~     (p=0.151 n=5+5)
Phases/tpcc-stock-level/Explore     148µs ± 1%   154µs ± 1%  +4.19%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/ExecBuild   157µs ± 1%   162µs ± 2%  +3.26%  (p=0.008 n=5+5)
Exec/tpcc-stock-level/ExecPlan      457µs ± 2%   447µs ± 2%  -2.11%  (p=0.032 n=5+5)
Exec/tpcc-stock-level/ExecOpt       437µs ± 1%   420µs ± 2%  -3.80%  (p=0.008 n=5+5)
```

Release note: None