@RaduBerinde 2018-09-20
Yeah, the filter pushdown rules don't apply to MergeJoin and LookupJoin (these only exist during exploration). If anything needs to use filter FDs, the filter FDs need to be augmented with the op-specific equalities (this is what the logical props/stat code does, via `joinPropsHelper`).